Extract functions from index to general helpers
